pyproject.toml
[project]
name = "kafka-cctv-streaming"
version = "0.1.0"
description = "kafka cctv streaming"
authors = [
    {name = "seilylook",email = "seilylook@naver.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "kafka-python (>=2.2.4,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "avro-python3 (>=1.10.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic (>=2.11.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.15.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"